Implementacja z użyciem IHttpActionResult

+++ Intel.DonutMaster/Intel.DonutMaster.Service/Controllers/ProductsController.cs
@@ -4,6 +4,7 @@ using Intel.DonutMaster.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Http;
 using System.Web;
 using System.Web.Http;
 
@@ -14,30 +15,67 @@ namespace Intel.DonutMaster.Service.Controllers
 
         private IProductsService _ProductsService = new MockProductsService();
 
-        public IList<Product> Get()
+        public IHttpActionResult Get()
         {
-            return _ProductsService.Get();
+            var products = _ProductsService.Get();
+
+            return Ok(products);
         }
 
-        public Product Get(int id)
+        public IHttpActionResult Get(int id)
         {
-            return _ProductsService.Get(id);
+            var product = _ProductsService.Get(id);
+
+            if (product==null)
+            {
+                return NotFound();
+            }
+
+            return Ok(product);
         }
 
 
-        public void Post(Product product)
+        public IHttpActionResult Post(Product product)
         {
             _ProductsService.Add(product);
+
+            // return StatusCode(System.Net.HttpStatusCode.Created);
+
+            return CreatedAtRoute("DefaultApi", new { id = product.ProductId}, product);
+
+            // return Created($"http://localhost:52765/api/products/{product.ProductId}", product);
         }
 
-        public void Put(Product product)
+        public IHttpActionResult Put(int id, Product product)
         {
+            if (id != product.ProductId)
+            {
+                return BadRequest("Niezgodny identyfikator");
+            }
+
             _ProductsService.Update(product);
+
+            return StatusCode(System.Net.HttpStatusCode.NoContent);
         }
 
-        public void Delete(int id)
+        public IHttpActionResult Delete(int id)
         {
+
+            if (id == 0)
+            {
+                return StatusCode(System.Net.HttpStatusCode.Conflict);
+
+                //var response = new HttpResponseMessage
+                //{
+                //    StatusCode = System.Net.HttpStatusCode.Conflict
+                //};
+
+                //return response;
+            }
+
             _ProductsService.Remove(id);
+
+            return Ok();
         }
 
         public void Head()
